requirements.txt
Creoleparser==0.7.5
Django==1.10
Genshi==0.7.7
Pillow==4.3.0
Pygments==2.5.2
django-taggit==0.22
https://github.com/oddstr13-openshell-no/django-taggit-templatetags/archive/90d02b4.zip#egg=django-taggit-templatetags
django-templatetag-sugar==1.0
gunicorn==18.0
#psycopg2==2.5.1
pytz==2014.2
sorl-thumbnail==12.4.1
wsgiref==0.1.2
python-dotenv==0.18.0
